# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KumbhSathi API - Pilgrimage companion for Kumbh Mela 2027
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register with email/password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /alerts:
    get:
      operationId: listAlerts
      tags: [alerts]
      summary: List all alerts and announcements
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [emergency, weather, crowd, announcement, event]
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      operationId: createAlert
      tags: [alerts]
      summary: Create alert (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
  /alerts/{alertId}:
    put:
      operationId: updateAlert
      tags: [alerts]
      summary: Update alert (admin)
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Delete alert (admin)
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ─── AI ───────────────────────────────────────────────────────────────────────
  /ai/chat:
    post:
      operationId: sendAiMessage
      tags: [ai]
      summary: Send message to Kumbh AI Guide
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatRequest"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/conversations/{sessionId}:
    get:
      operationId: getAiConversation
      tags: [ai]
      summary: Get conversation history
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiMessage"
    delete:
      operationId: clearAiConversation
      tags: [ai]
      summary: Clear conversation history
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Cleared

  # ─── Admin ────────────────────────────────────────────────────────────────────
...
  /admin/lost-found/{reportId}/approve:
    put:
      operationId: approveLostFoundReport
      tags: [admin]
      summary: Approve lost/found report
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LostFoundReport"
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  total:
                    type: integer
                required:
                  - users
                  - total

# ─── Schemas ──────────────────────────────────────────────────────────────────
components:
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        phone:
          type: string
        language:
          type: string
          enum: [en, hi, mr]
          default: hi
      required:
        - name
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    # User
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        avatar:
          type: string
          nullable: true
        language:
          type: string
          enum: [en, hi, mr]
        role:
          type: string
          enum: [user, admin, guest]
        emergencyContacts:
          type: array
          items:
            $ref: "#/components/schemas/EmergencyContact"
        medicalInfo:
          $ref: "#/components/schemas/MedicalInfo"
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - language
        - createdAt
...
    EmergencyContact:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        relation:
          type: string
      required:
        - id
        - name
        - phone
        - relation
    MedicalInfo:
      type: object
      properties:
        bloodGroup:
          type: string
        allergies:
          type: string
        conditions:
          type: string
        medications:
          type: string
...
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [emergency, weather, crowd, announcement, event]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, danger, critical]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - type
        - title
        - message
        - severity
        - isActive
        - createdAt
    CreateAlertRequest:
      type: object
      properties:
        type:
          type: string
          enum: [emergency, weather, crowd, announcement, event]
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, danger, critical]
        expiresAt:
          type: string
          format: date-time
      required:
        - type
        - title
        - message
        - severity
    UpdateAlertRequest:
      type: object
      properties:
        title:
          type: string
        message:
          type: string
        severity:
          type: string
          enum: [info, warning, danger, critical]
        isActive:
          type: boolean

    # Temples
...
    LostFoundReport:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [person, child, item]
        status:
          type: string
          enum: [open, searching, resolved, rejected]
        title:
          type: string
        description:
          type: string
        name:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        lastSeenLocation:
          type: string
        lastSeenTime:
          type: string
          format: date-time
        photos:
          type: array
          items:
            type: string
        contactName:
          type: string
        contactPhone:
          type: string
        reportedBy:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - status
        - title
        - description
        - lastSeenLocation
        - lastSeenTime
        - contactName
        - contactPhone
        - reportedBy
        - isApproved
        - createdAt
...
    AiChatRequest:
      type: object
      properties:
        message:
          type: string
        sessionId:
          type: string
        language:
          type: string
          enum: [en, hi, mr]
          default: hi
        context:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
            currentScreen:
              type: string
      required:
        - message
        - sessionId
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
        sessionId:
          type: string
        suggestions:
          type: array
          items:
            type: string
        relatedType:
          type: string
          nullable: true
          enum: [temple, event, service, emergency, null]
      required:
        - reply
        - sessionId
        - suggestions
    AiMessage:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - timestamp

    # Admin
...